+++ dear-imgui-sys/src/android.rs
@@ -1,8 +1,8 @@
-use ndk_sys::{AInputEvent, ANativeWindow};
+use std::ffi::c_void;
 
 unsafe extern "C" {
-    pub fn ImGui_ImplAndroid_Init(window: *mut ANativeWindow) -> bool;
-    pub fn ImGui_ImplAndroid_HandleInputEvent(input_event: *const AInputEvent) -> i32;
+    pub fn ImGui_ImplAndroid_Init(window: *mut c_void) -> bool;
+    pub fn ImGui_ImplAndroid_HandleInputEvent(input_event: *const c_void) -> i32;
     pub fn ImGui_ImplAndroid_Shutdown();
     pub fn ImGui_ImplAndroid_NewFrame();
 }